rename metadata, sync
