+++ ol-cli/src/node/mod.rs
@@ -3,7 +3,7 @@ pub mod node;
 pub mod account;
 pub mod chain_info;
 pub mod query;
-pub mod metadata;
+pub mod sync;
 pub mod client;
 pub mod states;
 mod transitions;
